requirements.txt
tensorflow_gpu==1.0.1
matplotlib==2.0.2
h5py==2.7.0
numpy==1.12.1
PyYAML==3.12
